Refactor to useForm hook
